Update system-order to Bevy 0.10.
This mostly involves removing references to labels, and pointing at system sets instead.
