+++ src/code/src/basics.rs
@@ -1614,94 +1614,6 @@ fn main() {
 }
 
 #[allow(dead_code)]
-mod app8a {
-use bevy::prelude::*;
-
-    fn particle_effects() {}
-    fn npc_behaviors() {}
-    fn enemy_movement() {}
-    fn player_movement() {}
-    fn input_handling() {}
-
-// ANCHOR: system-order
-fn main() {
-    App::new()
-        .add_plugins(DefaultPlugins)
-
-        // order doesn't matter for these systems:
-        .add_system(particle_effects)
-        .add_system(npc_behaviors)
-        .add_system(enemy_movement)
-
-        .add_system(input_handling)
-
-        .add_system(
-            player_movement
-                // `player_movement` must always run before `enemy_movement`
-                .before(enemy_movement)
-                // `player_movement` must always run after `input_handling`
-                .after(input_handling)
-        )
-        .run();
-}
-// ANCHOR_END: system-order
-}
-
-#[allow(dead_code)]
-mod app8b {
-use bevy::prelude::*;
-
-    fn input_joystick() {}
-    fn input_keyboard() {}
-    fn input_touch() {}
-    fn input_parameters() {}
-    fn player_movement() {}
-
-// ANCHOR: system-labels
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-#[derive(SystemLabel)]
-enum MyLabel {
-    /// everything that handles input
-    Input,
-    /// everything that updates player state
-    Player,
-    /// everything that moves things (works with transforms)
-    Movement,
-    /// systems that update the world map
-    Map,
-}
-
-fn main() {
-    App::new()
-        .add_plugins(DefaultPlugins)
-
-        // use labels, because we want to have multiple affected systems
-        .add_system(input_joystick.label(MyLabel::Input))
-        .add_system(input_keyboard.label(MyLabel::Input))
-        .add_system(input_touch.label(MyLabel::Input))
-
-        .add_system(input_parameters.before(MyLabel::Input))
-
-        .add_system(
-            player_movement
-                .before(MyLabel::Map)
-                .after(MyLabel::Input)
-                // we can have multiple labels on this system
-                .label(MyLabel::Player)
-                .label(MyLabel::Movement)
-                // can also use loose strings as labels
-                .label("player_movement")
-        )
-
-        // … and so on …
-
-        .run();
-}
-// ANCHOR_END: system-labels
-}
-
-#[allow(dead_code)]
 mod app9 {
 use bevy::prelude::*;
