+++ src/code010/src/lib.rs
@@ -24,6 +24,7 @@ mod d3 {
 }
 mod programming {
     mod intro_code;
+    mod system_order;
 }
 mod patterns {
 }
